+++ src/types/channel.d.ts
@@ -1,0 +1,16 @@
+type Channel = {
+  _id: string;
+  name: string;
+  description: string;
+  authRequired: boolean;
+  posts: string[];
+  createdAt: string;
+  updatedAt: string;
+  __v: number;
+};
+
+type ChannelItem = {
+  id: string;
+  name: string;
+  isSubscribe: boolean;
+};
